Renamed 'SharedStyledComponents' directory
