+++ src/components/RiskMapLegend/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import MapLegendLevel from "../MapLegendLevel";
-import { BlockDropshadow } from "../styled-components/Block";
+import { BlockDropshadow } from "../SharedStyledComponents/Block";
 import { MapLegendWrapper } from "./styles";
 
 const riskColors = ["#6c4ce1", "#7c64d3", "#9c8de7", "#c0b6fa", "#dad3fe"];
